+++ packages/radio/src/main.vue
@@ -1,40 +1,3 @@
-<template>
-  <label
-    class="wd-radio"
-    :class="{
-      'is-checked': isChecked,
-      'is-button': finalShape === 'button',
-      'is-disabled': finalDisabled,
-      'is-inline': inline
-    }"
-  >
-    <div
-      class="wd-radio__label"
-      :style="[{ 'color': isChecked && finalShape === 'button' && !finalDisabled ? finalCheckedColor : '' }, {'max-width': maxWidth}]"
-    >
-      <slot></slot>
-    </div>
-    <span
-      class="wd-radio__shape"
-      :style="{ 'color': isChecked && !finalDisabled ? finalCheckedColor : '' }"
-    >
-      <i
-        class="wd-icon-check"
-        :style="{
-          'color': isChecked && !finalDisabled ? finalCheckedColor : ''
-        }"
-      ></i>
-      <input
-        type="radio"
-        class="wd-radio__input"
-        :value="value"
-        :checked="isChecked"
-        :disabled="finalDisabled"
-        @change="handleChange"
-      />
-    </span>
-  </label>
-</template>
 
 <script>
 export default {
@@ -66,8 +29,76 @@ export default {
   },
   methods: {
     handleChange () {
+      console.log('修改')
       this.radioGroup.changeValue(this.value)
     }
+  },
+
+  render (h) {
+    const {
+      isChecked,
+      finalShape,
+      value,
+      finalDisabled,
+      finalCheckedColor,
+      maxWidth,
+      inline
+    } = this
+
+    const label = (
+      <div
+        class="wd-radio__label"
+        style={{
+          color: isChecked && finalShape === 'button' && !finalDisabled ? finalCheckedColor : '',
+          maxWidth: maxWidth
+        }}>{this.$slots.default}</div>
+    )
+    const input = (
+      <input
+        type="radio"
+        class="wd-radio__input"
+        value={value}
+        checked={isChecked}
+        disabled={finalDisabled}
+        onChange={this.handleChange}
+      />
+    )
+    const checkIcon = finalShape === 'check' ? (
+      <i class="wd-icon-check" style={{ color: isChecked && !finalDisabled ? finalCheckedColor : '' }}></i>
+    ) : ''
+    const shape = (
+      <span
+        class="wd-radio__shape"
+        style={{ color: isChecked && !finalDisabled ? finalCheckedColor : '' }}>
+        {checkIcon}{input}
+      </span>
+    )
+    // 单选框
+    const radio = inline ? (
+      <label
+        class={{
+          'wd-radio': true,
+          'is-checked': isChecked,
+          'is-button': finalShape === 'button',
+          'is-dot': finalShape === 'dot',
+          'is-disabled': finalDisabled,
+          'is-inline': inline
+        }}>
+        {shape}{label}
+      </label>
+    ) : (
+      <label
+        class={{
+          'wd-radio': true,
+          'is-checked': isChecked,
+          'is-button': finalShape === 'button',
+          'is-dot': finalShape === 'dot',
+          'is-disabled': finalDisabled
+        }}>
+        {label}{shape}
+      </label>
+    )
+    return radio
   }
 }
 </script>