+++ examples/docs/docs/changelog.md
@@ -2,7 +2,7 @@
 
 ### 2.0.0
 
-*2020-09-21*
+*2020-09-25*
 
 #### 新特性
 
@@ -17,6 +17,7 @@
 - Button
   - click 事件抛出event对象 (by [@yawuling](https://github.com/yawuling)） )
   - type 默认值改为 `primary` (by [@HXCStudio123](https://github.com/HXCStudio123) )
+  - round 默认值改为 `true` (by [@HXCStudio123](https://github.com/HXCStudio123) )
 - Card
   - 新增卡片组件 Card  (by [@awjing](https://github.com/awjing) )
 - Cell
@@ -102,11 +103,15 @@
   - 新增属性 `status` 设置进度条状态，支持sucess和danger (by [@awjing](https://github.com/awjing) )
 - Popover
   - 新增组件 `Popover` (by [@HXCStudio123](https://github.com/HXCStudio123) )
+- PullRefresh
+  - 新增插槽 `tip` (by [@HXCStudio123](https://github.com/HXCStudio123) )
 - Radio
   - 新增属性 `max-width`, 支持设置文字部分最大宽度，设置最大宽度开启文字折叠 (by [@HXCStudio123](https://github.com/HXCStudio123) )
+  - 属性 `type` 新增可选值 `check`，默认值变更为 `check` (by [@HXCStudio123](https://github.com/HXCStudio123) )
 - RadioGroup
   - 新增属性 `cell`, 支持单选组下的表单模式事件 (by [@HXCStudio123](https://github.com/HXCStudio123) )
   - 新增属性 `size`, 支持单选框尺寸更改（large） (by [@HXCStudio123](https://github.com/HXCStudio123) )
+  - 属性 `type` 新增可选值 `check`，默认值变更为 `check` (by [@HXCStudio123](https://github.com/HXCStudio123) )
 - SelectPicker
   - 新增组件 SelectPicker 。 (by [@HXCStudio123](https://github.com/HXCStudio123))
 - SortButton